# contracts/m0.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Runtime Platform — M0 Contract Lab",
    "version": "1.0.0-m0",
    "description": "Local-only schema validation and metadata persistence. NO provider calls, execution scheduling, budget ledger, or production authentication."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4311"
    }
  ],
  "security": [
    {
      "localBearer": []
    }
  ],
  "paths": {
...
    "/api/m0/validations": {
      "post": {
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent replay of prior validation"
          },
          "201": {
            "description": "Validation metadata persisted, not an AI execution"
          },
          "400": {
            "description": "Malformed lab envelope"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "409": {
            "description": "Same key, different request digest"
          },
          "503": {
            "description": "Persistence unavailable; no successful save claimed"
          }
        }
      }
    },
    "/api/m0/history": {
      "get": {
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful local contract-lab response"
          }
        }
      }
    },
    "/api/m0/history/{id}": {
      "get": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful local contract-lab response"
          },
          "404": {
            "description": "Record absent or belongs to another app"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "localBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Generated local app credential; BFF injects it, never exposed to browser."
      }
    },
    "schemas": {
...
      "ValidationInput": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "chat",
              "generate",
              "execution"
            ]
          },
          "payload": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "kind",
          "payload"
        ],
        "additionalProperties": false
      }
    }
  }
}